rename options package
