+++ internal/pkg/validators/email_address_validator_test.go
@@ -1,4 +1,4 @@
-package options
+package validators
 
 import (
 	"testing"
