+++ internal/pkg/validators/email_domain_validator.go
@@ -1,4 +1,4 @@
-package options
+package validators
 
 import (
 	"errors"
