+++ internal/pkg/validators/validators.go
@@ -1,4 +1,4 @@
-package options
+package validators
 
 import (
 	"errors"
